\name{BOOOM-package}
\alias{BOOOM-package}
\alias{BOOOM}
\docType{package}
\title{
  BOOOM
}
\description{
  Coming Soon.
}
\details{
  Coming Soon.
}
\author{
Priyam Das.

Maintainer: Priyam Das <pdas@ncsu.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
